removed 'local' flag from BGSynth stuff, combined BGR with BGRworker.js.
