+++ src/BGRworker.js
@@ -1,7 +1,185 @@
 import * as Comlink from "comlink";
+import HydraSynth from "./hydra-synth.js";
+import {BGHydraSynth} from "./BGHydra-synth.js";
 
-import {BGR} from './BGR.js';
+import {Deglobalize} from './Deglobalize.js';
 
-let BGRWorker = BGR;
+const GeneratorFunction = function* () {}.constructor;
 
-Comlink.expose(BGRWorker);
+
+class BGRworker {
+  constructor(canvas, useWGSL=false) {
+    if (!(typeof self !== "undefined" && self.constructor && self.constructor.name === "DedicatedWorkerGlobalScope")) {
+    	this.isWebWorker = false;
+		} else {
+    	this.isWebWorker = true;
+		}
+		this.can = canvas;
+		this.directToCanvas = this.can ? true : false;
+		this.useWGSL = useWGSL;
+  } 
+
+	destroy() {
+		if (this._h) {
+			this._h._destroy();
+		}
+	}
+		
+
+  registerCallback(name, cb) {
+		if (name === 'frame') {
+			this.frameCB = cb;
+		} else if (name === 'proxy') {
+			this.proxyCB = cb;
+		}
+	}
+
+
+	async openHydra() {
+			if (this.h === undefined) {
+			if (!this.directToCanvas) {
+				this.can = new OffscreenCanvas(1920, 1080);
+		  }
+		  let hs = new BGHydraSynth({useWGSL: this.useWGSL,  worker: this, makeGlobal: false, canvas: this.can,  autoloop: false, detectAudio: false, enableStreamCapture: false });
+		  if (hs.wgslPromise) await hs.wgslPromise;
+    	this.h = hs.synth;
+
+    	console.log("BGHydraSynth created: " + this.h);
+    }
+	}
+
+
+  async setSketch(inStr) {
+  	if (!this.h) return;
+  	let str;
+  	let errFound = "";
+  	try {
+  		str = Deglobalize(inStr, "_h");
+  	} catch (err) {
+  		errFound = err;
+  		console.log("Deglobalize error " + err);
+  		str = inStr;
+  	}
+  	let keys = Object.keys(this.h);
+    let values = [];
+    for (let i = 0; i < keys.length; ++i) values.push(this.h[keys[i]]);
+    keys.push("h");
+    values.push(this.h);
+    keys.push("_h"); // _h used for fixing-up primitive-valued 'global' references, like "time".
+    values.push(this.h);
+    try {
+    	let fn = new GeneratorFunction(...keys, str);
+    	this.h.generator = fn(...values);
+    } catch (err) {
+    		console.log("Error compiling generator function");
+    		console.log(err);
+    		this.stopEarlierTimer();
+    		return;
+    }
+    this.stopEarlierTimer();
+    try {
+    	let reply = this.h.generator.next();
+    	this.planNext(reply);
+    } catch (err) {
+    	console.log("Error calling initial generator function.next()");
+    	console.log(err);
+    	delete this.h.generator;
+    	return;
+    }
+}
+
+  stopEarlierTimer() {
+	  if(this.timeOutKey !== undefined) {
+		  clearTimeout(this.timeOutKey);
+		  this.timeOutKey = undefined;
+	  }
+}
+
+ generatorTick() {
+	if (!this.h || !this.h.generator) return;
+	try {
+		let f = this.h.generator;
+		let reply = f.next();
+		this.planNext(reply);
+	} catch (err) {
+    	console.log("Error calling generator function.next()");
+    	console.log(err);
+    	delete this.h.generator;
+	}
+}
+
+ planNext(reply) {
+	 if (!reply) return;
+
+   if (!reply.done) {
+    		let wT = reply.value;
+    		if (wT === undefined) {
+    			wT = 10;
+    		} else {
+    			wT = wT * 1000; // Convert to ms.
+    		}
+    		let that = this;
+    		this.timeOutKey = setTimeout(()=>that.generatorTick(), wT);
+    } else {
+    		delete this.h.generator;
+    		this.timeOutKey = undefined;
+    }
+}
+  async hush() {
+  	 if (!this.h) return;
+  	 this.h.hush();
+  }
+
+	async tick(dt, mouseData, timeV) {
+		if (this.h) {
+				if (mouseData && this.isWebWorker) {
+					this.h.mouse.x = mouseData.x;
+					this.h.mouse.y = mouseData.y;
+				}
+				//this.h.time+= (dt / 1000.0);
+				if (this.isWebWorker && this.directToCanvas) {
+					// direct to canvas sends nothing back via the frameCB.
+					// It just signals the next frame was processed.
+					await this.h.tick(dt);
+					if (this.frameCB) this.frameCB();
+				} else {
+				 this.h.tick(dt);
+				if (this.frameCB) {
+					let fr = this.can.transferToImageBitmap();
+					if (this.isWebWorker) {
+						this.frameCB(Comlink.transfer(fr, [fr]));
+					} else {
+						this.frameCB(fr);
+					}
+				}
+			}
+		}
+	}
+
+	getFrameData() {
+		let img = this.can.transferToImageBitmap();
+		return img;
+	}
+
+	async openSourceProxy(kind, sourceX, mediaAddr, params) {
+		// Forward open proxy request via proxy callback to the HydraStage
+		if (this.proxyCB) {
+			this.proxyCB(kind, sourceX, mediaAddr, params);
+		} else {
+			console.log("No proxy callback registered.");
+		}
+	}
+
+	async proxyFrameUpdate(sourceX, img) {
+		let h = this.h;
+		if (h) {
+			let sName = 's' + sourceX;
+			let st = h[sName];
+			st.injectImage(img);
+		} else {
+			console.log("No hydra to update in BGRWorker");
+		}
+	}
+}
+Comlink.expose(BGRworker);
+export {BGRworker}
